Rename util class and setValues method
