+++ src/main/java/org/kainos/ea/db/DaoUtil.java
@@ -1,14 +1,15 @@
-package org.kainos.ea;
+package org.kainos.ea.db;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
-public final class Util {
-    private Util() {
+public final class DaoUtil {
+    private DaoUtil() {
 
     }
 
-    public static void setValues(PreparedStatement preparedStatement, Object... values)
+    public static void populatePreparedStatement(
+            PreparedStatement preparedStatement, Object... values)
             throws SQLException {
         for (int i = 0; i < values.length; i++) {
             preparedStatement.setObject(i + 1, values[i]);
